pyproject.toml
[project]
name = "talking-avatar"
version = "0.1.0"
description = "Create AI-powered talking avatars with ease"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.5",
    "uvicorn[standard]>=0.34.0",
    "python-dotenv>=1.0.1",
    "fal-client>=0.5.6",
    "python-multipart>=0.0.18",
    "requests>=2.32.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "httpx>=0.24.0",
]